+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 ### 2023-10-09
+ - Input parameters have new constraint. If a user supplies a scaleExtent or
+   scaleSize in the request, the target CRS must also be included or the
+   request will fail.
  - Output dimension calculation no based on input data resolution and the scale
    extent.